Input Component Done
