+++ frontend/client/src/components/Input.tsx
@@ -1,62 +1,80 @@
-import { ChangeEvent, ReactNode } from "react";
-import { InfoIcon } from "./icons";
+import { useState, ChangeEvent } from "react";
+import { CheckMark, InfoIcon } from "./icons";
+import { InputProps } from "../types";
 
-interface InputProps {
-  label?: string;
-  placeholder?: string;
-  id?: string;
-  type?: "text" | "email" | "password";
-  name?: string;
-  value?: string;
-  errorText?: string;
-  successText?: string;
-  className?: string;
-  labelClassName?: string;
-  inputClassName?: string;
-  icon?: ReactNode;
-  iconClassName?: string;
-
-  onChange?: (event: ChangeEvent<HTMLInputElement>) => void;
-}
 const Input = ({
-  className = "",
-  label = "",
-  id = "",
-  placeholder = " ",
-  labelClassName = "",
-  inputClassName = "",
   icon,
-  iconClassName,
+  name = "",
   type = "text",
-  errorText,
-  successText,
+  label = "",
+  value = "",
+  register,
+  onChange,
+  className = "",
+  errorText = "",
+  successText = "",
+  placeholder = "",
+  labelClassName = "",
 }: InputProps) => {
+  const [inputValue, setInputValue] = useState(value || "");
+
+  const handleInputChange = (e: ChangeEvent<HTMLInputElement>) => {
+    setInputValue(e.target.value);
+    if (onChange) onChange(e);
+  };
+
   return (
-    <div className={`font-metropolis relative border ${className}`}>
-      {label && (
-        <label
-          htmlFor={id || ""}
-          className={`text-lg label font-semibold ${labelClassName}`}
-        >
-          {label} :
-        </label>
+    <div className="w-[400px] space-y-1">
+      <div className="relative">
+        {/* Input */}
+        <input
+          type={type}
+          id={name}
+          name={name}
+          value={inputValue}
+          onChange={handleInputChange}
+          placeholder={placeholder ? placeholder : " "}
+          className={`peer w-full focus:outline-none flex-grow outline-none font-normal text-base overflow-hidden bg-smoke-eerie rounded-xl border border-primary-inverted-10 p-4 text-primary-inverted ${
+            icon && "pr-10"
+          } ${className}`}
+          {...register}
+        />
+        {/* Floating Label */}
+        {label && (
+          <label
+            htmlFor={name}
+            className={`absolute left-5 border text-primary-inverted ${
+              inputValue
+                ? "top-0 py-1 px-3 border-primary-inverted-10 text-sm rounded-full bg-smoke-eerie"
+                : "top-1/2 border-transparent"
+            } transform -translate-y-1/2 transition-all duration-200 leading-none peer-focus:top-0 peer-focus:text-sm cursor-text peer-focus:leading-none peer-focus:px-3 peer-focus:py-1 peer-focus:border-primary-inverted-10 peer-focus:bg-smoke-eerie peer-focus:rounded-full ${labelClassName}`}
+          >
+            {label}
+          </label>
+        )}
+
+        {/* Icon */}
+        {icon && (
+          <span className="h-full absolute top-0 right-0 p-3 flex justify-center items-center">
+            {icon}
+          </span>
+        )}
+      </div>
+
+      {/* Error Message */}
+      {errorText && !successText && (
+        <span className="flex gap-1 items-center text-red-500 text-xs font-normal mt-2">
+          <InfoIcon className="fill-red-500" />
+          {errorText}
+        </span>
       )}
-      <input
-        type={type}
-        placeholder={placeholder}
-        className={`focus:outline-none flex-grow outline-none font-normal text-base overflow-hidden ${inputClassName}`}
-      />
-      {icon && (
-        <div
-          className={`absolute right-2 top-1/2 -translate-y-1/2 ${iconClassName}`}
-        >
-          {icon}
-        </div>
+      {/* Success Message */}
+      {successText && !errorText && (
+        <span className="flex gap-1 items-center text-green-500 text-xs font-normal">
+          <CheckMark className="fill-green-500" />
+          {successText}
+        </span>
       )}
-      {errorText && !successText &&<span className="flex gap-1 items-center text-red-500 text-xs font-normal mt-2">
-        <InfoIcon className="fill-red-500" />
-        {errorText}
-      </span>}
     </div>
   );
 };
